feat: made getUrl function as hook and move country dependencies inside
